first_commit
Through the K-Nearest Neighbor library, we feed the normalized unknown_points which represent 3 separate parameters that help predict if the movie will be good or bad (1 or 0). The unknown labels stores the actual results for the validation set. 

There's only one possible combination that yields 100% in the dataset (which is if all values of the variables -  y_predicted and unknown_labels - are equal. 

There are 3 possible combinations to yield 67% in the dataset (which is given in line 24). 

There are 3 possible combinations to yield 1/3 (which is given in line 26. 

There is 1 possible combination to yield 0/3 (which is given in line 28). 

Throught his we can optimize for the best K (by reading the true_false_list). Don't overfit but also don't underfit. Fit it as necessary lol.