Create and Upload own k_nearest algo
The function, "distance" identifies the Euclidean Distance between 2 separate lists (i.e. movie1, and movie2). It takes a look at the length of the movie and runs a for loop on it's range where it identifies the squared_distance between each item in the list. For instance [0, 1, 2] and [3, 4, 5] (not normalized) would yield (0-3)^2 + (1-4)^2 + (2-5)^2. From there it square roots the result to identify the final distance. 

The classification algorithm takes in an unknown dataset of type list that looks something like this: [[a, b, c], [d, e, f]...[x, y, z]]. It takes in a dataset to train itself (i.e. "Hello": [a, b, c], "Shutter Island": [d, e, f], "Inception": [g, h, k]). It takes in labels (which is -- "Tenet": [1],  "Twilight": [0] ... --- good or bad). It takes in a particular k as well. 

Furthermore, after creating an empty distances list, it runs through each title in the dataset. It identifies the distance_to_point between each movie in "dataset" and "unknown." It appends the distance_to_point variable to the list distances with the title. Righg after, it sorts it and slices to the right k in the variable neighbors. 

Now the algorithm will analyze if the movie is good or bad (1 or 0). For each distance in neighbors we look at the distance_to_point variable (which again, lists how far our unknown movie is to the "neighbor" in neighbors. This data is stored in title. Now it identifies each of the title of neighbors in the labels to check if it's good or bad. It increases num_bad or num_good depending on if it's good or not (to identify if the passed unknown movie is good or bad). 

Lastly, we measure the validation accuracy through the function -- find_validation_accuracy. Simply, it loops through movies. If the guess is correcct then it increases num_correct by 1. To measure accuracy, we take the num_correct divide it by validation_set.